fix: ページを3つに分割してスクロールしやすくし、移動を4回行って寄せていくようにした
